Attempt to patch bug when using spectrometers for veryshort time.

+++ spectro.py
@@ -574,14 +574,14 @@ class AvaSpec_Handler:
 
         # Init c_MeasConfigType to pass it to AVS_PrepareMeasure.
         Meas = c_MeasConfigType()
-        Meas.m_StartPixel = 0
-        Meas.m_StopPixel = numPix.value - 1  # Last pixel.
-        Meas.m_IntegrationTime = intTime
-        Meas.m_NrAverages = nrAverages if triggerred else 1
+        Meas.m_StartPixel = ctypes.c_int(0)
+        Meas.m_StopPixel = ctypes.c_int(numPix.value - 1)  # Last pixel.
+        Meas.m_IntegrationTime = ctypes.c_int(intTime)
+        Meas.m_NrAverages = ctypes.c_int(nrAverages if triggerred else 1)
         # Trigger configuration.
-        Meas.m_Trigger.m_Mode = int(triggerred)
-        Meas.m_Trigger.m_Source = 0
-        Meas.m_Trigger.m_SourceType = 0
+        Meas.m_Trigger.m_Mode = ctypes.c_int(int(triggerred))
+        Meas.m_Trigger.m_Source = ctypes.c_int(0)
+        Meas.m_Trigger.m_SourceType = ctypes.c_int(0)
 
         AVS_DLL.AVS_PrepareMeasure(device, ctypes.byref(Meas))
